+++ restmachine/__init__.py
@@ -14,6 +14,7 @@ from .content_renderers import (
     JSONRenderer,
     PlainTextRenderer,
 )
+from .drivers import AwsApiGatewayDriver, Driver
 from .exceptions import ValidationError
 from .models import HTTPMethod, Request, Response
 
@@ -31,4 +32,6 @@ __all__ = [
     "PlainTextRenderer",
     "ContentRenderer",
     "ValidationError",
+    "Driver",
+    "AwsApiGatewayDriver",
 ]